fix
теперь чистый класс
add невозможность ввести пустые данные, null рабботает как отмена. нельзя запросить несуществующие данные. список может быть пустым

+++ HomeWorks/N.13 DRINKS_HASH_FUNC/drinks_hash_func.js
@@ -6,7 +6,6 @@ function HashStorageFunc() {
         storage[key] = value;
         console.log(storage);
     }
-
     this.getValue = (key) => {
         return storage[key];
     }
@@ -16,10 +15,8 @@ function HashStorageFunc() {
     this.deleteValue = (delKey) => {
         if (delKey in storage) {
             delete storage[delKey];
-            console.warn(`Напиток ${delKey} удалён!`);
             return true;
         } else {
-            console.error(`Напиток ${delKey} не  был передан!`);
             return false;
         }
     }
@@ -28,22 +25,45 @@ function HashStorageFunc() {
 var drinkStorage = new HashStorageFunc();
 // 3
 addButton.onclick = () => {
-    var drinkName = prompt('Введите название напитка', '');
+    do {
+        var drinkName = prompt('Введите название напитка', '');
+        if (drinkName === null) return false;
+        if (!drinkName) console.error(`Название напитка не может быть пустым!`);
+    } while (!drinkName);
     var alco = confirm(`Алкогольный? `);
     var recept = prompt(`Введите рецепт... `, '');
-    // return
     drinkStorage.addValue(drinkName, { a: alco, r: recept })
-
 }
 getButton.onclick = () => {
-    var drinkName = prompt(`Введите название напитка ...`, '');
+    do {
+        var drinkName = prompt(`Введите название напитка ...`, '');
+        if (drinkName === null) return false;
+        if (!drinkName) console.error(`Название напитка не может быть пустым!`);
+    } while (!drinkName);
     var i = drinkStorage.getValue(drinkName);
-    console.info(`Напиток: ${drinkName}\nАлкогольный: ${i.a ? 'Да' : 'Нет'}\nРецепт: ${i.r}`)
+    if (i) {
+        console.info(`Напиток: ${drinkName}\nАлкогольный: ${i.a ? 'Да' : 'Нет'}\nРецепт: ${i.r}`)
+    } else {
+        console.error(`Напиток "${drinkName}" ещё не  был внесен!`);
+    }
 }
 deleteButton.onclick = () => {
-    var delKey = prompt('Введите название напитка для удаления', '');
-    drinkStorage.deleteValue(delKey);
+    do {
+        var delKey = prompt('Введите название напитка для удаления', '');
+        if (delKey === null) return false;
+        if (!delKey) console.error(`Название удаляемого напитка не может быть пустым!`);
+    } while (!delKey);
+    if (drinkStorage.getValue(delKey)) {
+        drinkStorage.deleteValue(delKey);
+        console.warn(`Напиток "${delKey}" удалён!`);
+    } else {
+        console.error(`Напитка "${delKey}" нет в списке!`);
+    }
 }
 listButton.onclick = () => {
-    console.info(drinkStorage.getKeys().join('\n'));
+    if (drinkStorage.getKeys() == false) {
+        console.error(`Список напитков пуст!`);
+    } else {
+        console.info(drinkStorage.getKeys().join('\n'));
+    }
 }
